% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_related_genes}
\alias{covid_related_genes}
\title{COVID-19 infection related genes}
\format{
\subsection{\code{covid_related_genes}}{

A character vector of 57 gene IDs
}
}
\usage{
covid_related_genes
}
\description{
HGNC gene IDs
}
\keyword{datasets}
