% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{indra_discrete_gsea}
\alias{indra_discrete_gsea}
\title{Perform discrete gene set enrichment analysis using INDRA}
\usage{
indra_discrete_gsea(
  genes,
  method = "fdr_bh",
  alpha = 0.05,
  keep_insignificant = FALSE,
  minimum_evidence_count = 2,
  minimum_belief = 0,
  indra_path_analysis = TRUE,
  background_genes = NULL
)
}
\arguments{
\item{genes}{List of gene identifiers (HGNC symbols or identifiers)}

\item{method}{Statistical method to apply)}

\item{alpha}{Significance level}

\item{keep_insignificant}{Whether to retain insignificant results}

\item{minimum_evidence_count}{Minimum number of evidence for inclusion}

\item{minimum_belief}{Minimum belief score for filtering}

\item{indra_path_analysis}{Whether to perform INDRA pathway analysis}

\item{background_genes}{List of background genes}
}
\value{
A list containing results per analysis type as data frames
}
\description{
Perform discrete gene set enrichment analysis using INDRA
}
