% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{indra_ranked_analysis}
\alias{indra_ranked_analysis}
\title{Perform ranked gene set analysis}
\usage{
indra_ranked_analysis(
  gene_names,
  log_fold_change,
  species = c("rat", "mouse", "human"),
  permutations = 100,
  source = c("go", "reactome", "wikipathways", "phenotype", "indra-upstream",
    "indra-downstream"),
  alpha = 0.05,
  keep_insignificant = FALSE,
  minimum_evidence_count = 2,
  minimum_belief = 0
)
}
\arguments{
\item{gene_names}{Vector of gene names}

\item{log_fold_change}{Vector of log fold changes}

\item{species}{Species ('rat', 'mouse', or 'human')}

\item{permutations}{Number of permutations}

\item{source}{Analysis type ('go', 'reactome', 'wikipathways', 'phenotype', 'indra-upstream', 'indra-downstream')}

\item{alpha}{Significance level}

\item{keep_insignificant}{Whether to retain insignificant results}

\item{minimum_evidence_count}{Minimum number of evidence for inclusion}

\item{minimum_belief}{Minimum belief score for filtering}
}
\description{
Perform ranked gene set analysis
}
